pyproject.toml
[project]
name = "ogame-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "playwright>=1.57.0",
    "python-dotenv>=1.2.1",
]
